Check whether filter list is a sequence type explicitly
Add an explicit PySequence_Check() for filter list processing instead of
relying on PySequence_Length() raising a TypeError for dict. This fixes
support for PyPy which successfully returns the dict length.

The CPython behavior seems to a bug, as the documentation clearly
states that the call on non-sequence type is equivalent to len(). It
seem more prudent therefore to check for sequence type explicitly.

Closes: https://github.com/peterjc/backports.lzma/issues/5