Remove the setuptools namespace package declaration in setup.py.
All of this, both __init__.py and setup.py should be using python
standard namespace package technology, not setuptools namespace
technology.  If we attempt to use setuptools technology we'll break all
the other packages that are using the backpors namespace.

This is the real cause of https://github.com/peterjc/backports.lzma/issues/8 and
https://github.com/peterjc/backports.lzma/issues/16

Fixes #28
Fixes #16
Fixes #8